requirements.txt
streamlit
anthropic
python-dotenv
fpdf2
requests